+++ packages/react/index.ts
@@ -1,20 +1,24 @@
-import { useReducer, useEffect, useLayoutEffect } from "react";
-import { manageStore } from "@reffect/core";
+import { useEffect, useLayoutEffect, useState } from "react";
+import { manage, EffectState, Action } from "@reffect/core";
 
 const useIsomorphicEffect = typeof window === "undefined" ? useEffect : useLayoutEffect;
 
-const reducer = (state: any, newState: any) => ({ ...newState });
-
 export const useStore = <Store extends object>(store: Store): Store => {
-  const [state, triggerUpdate] = useReducer(reducer, store);
+  const [state, setState] = useState<Store>(store);
 
-  useIsomorphicEffect(() => {
-    const subscriber = () => triggerUpdate(store);
-    const { subscribe, unsubscribe } = manageStore(store);
-
-    subscribe(subscriber);
-    return () => unsubscribe(subscriber);
-  }, []);
+  useIsomorphicEffect(() => manage(store).subscribe(() => setState(store)), []);
 
   return state;
+};
+
+export const useEffectState = (effect: Action<unknown[], unknown>) => {
+  const [state, setState] = useState<EffectState | null>(null);
+
+  useIsomorphicEffect(() => manage(effect).subscribe(state => setState(state)), []);
+
+  return {
+    pending: state === EffectState.Pending,
+    fail: state === EffectState.Fail,
+    done: state === EffectState.Done,
+  };
 };