+++ packages/localStore/CHANGELOG.md
@@ -1,5 +1,11 @@
+# 1.0.6
+
+### Minor:
+
+1. Support new `@reffect/core` version (`1.4.0`)
+
 # 1.0.4 - 1.0.5
 
-Chore:
+### Minor:
 
 1. Support new `@reffect/core` version (`1.2.0`)