+++ packages/react/CHANGELOG.md
@@ -1,5 +1,15 @@
+# 1.1.0
+
+### Features:
+
+1. Add `useEffectState(effect)` for watch effect's state
+
+### Minor:
+
+1. Support new `@reffect/core` version (`1.4.0`)
+
 # 1.0.3 - 1.0.4
 
-Chore:
+### Minor:
 
 1. Support new `@reffect/core` version (`1.2.0`)